+++ src/singleTable/model/from/fromEntity/definitions/crud.ts
@@ -5,12 +5,18 @@ import type {
   UpdateParams,
   DeleteItemParams,
 } from 'provider/utils';
+import { FirstParameter } from 'types';
 
-export type EntityGetParams<Registered extends ExtendableSingleTableEntity> = Parameters<
+type NoKeyParam<Registered extends ExtendableSingleTableEntity> = FirstParameter<
   Registered['getKey']
->[0] extends undefined
-  ? [Omit<GetItemParams<Registered['__entity']>, 'table' | 'key'>?]
-  : [Omit<GetItemParams<Registered['__entity']>, 'table' | 'key'> & KeyParams<Registered>];
+> extends undefined
+  ? true
+  : false;
+
+export type EntityGetParams<Registered extends ExtendableSingleTableEntity> =
+  NoKeyParam<Registered> extends true
+    ? [Omit<GetItemParams<Registered['__entity']>, 'table' | 'key'>?]
+    : [Omit<GetItemParams<Registered['__entity']>, 'table' | 'key'> & KeyParams<Registered>];
 
 export type EntityBatchGetParams<Registered extends ExtendableSingleTableEntity> = Omit<
   BatchListItemsArgs<Registered['__entity']>,
@@ -20,7 +26,9 @@ export type EntityBatchGetParams<Registered extends ExtendableSingleTableEntity>
 };
 
 export type DeleteEntityParams<Registered extends ExtendableSingleTableEntity> =
-  KeyParams<Registered> & Omit<DeleteItemParams<Registered['__entity']>, 'table' | 'key'>;
+  NoKeyParam<Registered> extends true
+    ? [Omit<DeleteItemParams<Registered['__entity']>, 'table' | 'key'>?]
+    : [KeyParams<Registered> & Omit<DeleteItemParams<Registered['__entity']>, 'table' | 'key'>];
 
 export type UpdateEntityParams<Registered extends ExtendableSingleTableEntity> = Omit<
   UpdateParams<Registered['__entity']>,